maths-note-be/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pillow
python-dotenv
groq
